new parallel fixes some issues

+++ src/utils.c
@@ -5,5 +5,5 @@
 unsigned long cur_time_millis(void) {
     struct timeval timeval;
     gettimeofday(&timeval, NULL);
-    return (long) timeval.tv_sec * 1000 + (timeval.tv_usec / 1000);
+    return (long) (long)timeval.tv_sec * 1000 + (timeval.tv_usec / 1000);
 }
